Move templates folder to js > components
