+++ resources/js/components/templates/TemplateSelectCard.vue
