--- resources/js/processes/templates/index.js
@@ -1,4 +1,0 @@
-import Vue from "vue";
-import TemplateExistsModal from "./components/TemplateExistsModal.vue";
-import TemplateSelect from "./components/TemplateSelect.vue";
-import TemplateSelectCard from "./components/TemplateSelectCard.vue";
